Refactor growth form state logic
